Huge fixes. Database is now populated and clicking on a business name shows the correct info on the card.

+++ app/src/main/java/edu/uark/fayettefinds/AddEditBusinessCard/AddEditBusinessCardViewModel.kt
@@ -14,7 +14,7 @@ class AddEditBusinessCardViewModel(private val repository: FayetteFindsRepositor
     val businessCard:LiveData<BusinessCard>
         get() = _businessCard
 
-    fun start(itemId:Int)
+    fun start(itemId:Long)
     {
         viewModelScope.launch {
             repository.allBusinessCards.collect{
